+++ documentation/Changelog.md
@@ -179,6 +179,7 @@ Time Copy was converted into a side-bar version from version 0.8.7.
 - fix: showAllMessages default value
 - fix: other import/export issues
 - fix: create lstorasge on startup
+- fix: dlc amag protime - catch error when overlay is not hiding
 - other: restructure and clean up code
 - other: map theme select dynamicaly
 - other: profile cleanup (redesign and select options)